fix(satmae,satmaepp): drop center-crop preprocessing, fix SatMAE input scaling
Align SatMAE/SatMAE++ RGB preprocessing with the ScaleMAE direct-resize fix so
the ViT patch-token grid covers the full fetched square, keeping the ROI
crop-back and tile stitcher geometrically aligned.

- satmae: stop routing through the rshf wrapper's transform() (which does
  Resize(256)+CenterCrop(224) and, when handed a float32 array, also skips
  ToTensor's /255 — driving inputs ~255x out of distribution and producing
  noise). Replace with an explicit Resize((image_size, image_size)) + ToTensor
  + Normalize using the model config's fMoW mean/std (fallback to constants),
  no center crop. Record preprocess provenance in meta on all three paths.
- satmaepp: replace Resize(short=256)+CenterCrop(224) with a direct
  Resize((image_size, image_size)); drop the now-unused short-side helper and
  update preprocess_info (resize_to / center_crop=None).
- remoteclip: pass uint8 (not float32) into the wrapper transform so a
  ToTensor()-based pipeline scales to [0,1] (latent; that branch is currently
  dead since rshf RemoteCLIP exposes .preprocess, not .transform).

Full test suite passes (880 passed, 2 skipped).

+++ src/rs_embed/embedders/onthefly_remoteclip.py
@@ -352,7 +352,9 @@ def _remoteclip_encode_tokens(
 
     # --- preprocess to tensor ---
     if hasattr(model, "transform") and callable(model.transform):
-        x = model.transform(rgb_u8.astype(np.float32), image_size).unsqueeze(0)
+        # Pass uint8: a transform built on ToTensor() only scales uint8 to [0,1];
+        # a float32 array skips that /255 and Normalize() then runs ~255x OOD.
+        x = model.transform(rgb_u8, image_size).unsqueeze(0)
     else:
         preprocess = transforms.Compose(
             [
@@ -483,7 +485,9 @@ def _remoteclip_encode_pooled_batch(
     xs = []
     if hasattr(model, "transform") and callable(model.transform):
         for rgb_u8 in rgb_u8_batch:
-            x = model.transform(rgb_u8.astype(np.float32), image_size)
+            # uint8 in: ToTensor()-based transforms only divide by 255 for uint8;
+            # float32 skips it and Normalize() then runs ~255x out of distribution.
+            x = model.transform(rgb_u8, image_size)
             if not torch.is_tensor(x):
                 raise ModelError("RemoteCLIP transform did not return torch.Tensor.")
             if x.ndim != 3:
